[flow][refactor] Make the logic of CallLatentPredT completely internal to predicate_kit by using existing concretization primitives
Summary:
`CallLatentPredT` is at the concretization point of `ConcretizeForInspection`, so we can directly use it to concretize the `call_latent_pred`. The only thing we need to handle additionally is intersections, which now can be done easily with `Speculation_kit.try_custom`. As a result, we can have the logic of `CallLatentPredT` completely inside predicate_kit, which prepares for the eventual breaking of cyclic dependency between flow_js and predicate_kit.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D59944468

fbshipit-source-id: 4aaacadf199b64b4ec9a1ce571bec4c72688c332