[flow][refactor] Make `PredicateT`, `PreprocessKitT (SentinelPropTest)` and `PreprocessKitT (PropExistsTest)` only do concretization
Summary:
With all the preparation that's happening before this diff, it's finally clear that for `PredicateT`, `PreprocessKitT (SentinelPropTest)` and `PreprocessKitT (PropExistsTest)`, we are only using them for concretization purpose now. In Flow_js, they are handled like

```
| (OpaqueT _, PredicateT (p, t)) -> PredicateKit.predicate cx trace t l p
```

which we can change it to

```
| (OpaqueT _, PredicateT (p, t)) -> rec_flow_t cx trace ~use_op:unknown_use (l, OpenT tvar)
```

and on the receiving end of the tout tvar `t`, we can just get all possible types out of `t`.

Note that this is not safe before LTI, but now given the strong ordering guarantee of LTI, we definitely know that the input to the predicate use_ts are already resolved and won't get any new bounds, we can finally do it!

Therefore, this diff mostly did the mechanical work of turning these use_ts into a concretizer. The only interesting thing is for `PredicateT` in handle generic, we need to replicate the logic in predicate_kit to get the behavior of `PredicateT(GenericT(bound) = GenericT(bound=PredicateT(bound))`

With these changes done, predicate_kit and flow_js no longer have cyclic dependencies, so we can unfunctorize predicate_kit, but I will leave it to another diff to keep this one more focused.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D59949869

fbshipit-source-id: bb9e2e511caaa5ac2cfe767ce8c21366f96036b5