[flow] No more `InternalT (ExtendsT _)` 1/n: Introduce concretize_and_run_predicate to let us call relevant predicate runner directly
Summary:
We have a lot of `predicate` call here. For a lot of cases, we know exactly which predicate runner (e.g. instanceof, sentinel prop, etc) to run after concretization. This diff introduces this helper that's a generalization of the current `predicate` function, and migrate recursive calls in instanceof_test to use it.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D59979293

fbshipit-source-id: 7f4fb40ef8f3a2002d29b4e08ec62d4d258054e1