[ci] Fix windows build (#9193)
Summary:
We are now getting weird errors like `The msstore source requires that you view the following agreements before using.`. Not sure what's going on and what has changed, but  we can avoid it by not using winget to install opam.

The winget setup is the only thing I can find back then when opam 2.2 is still in beta during ocaml 5 migration, but now opam 2.2 is stable and in [this blog post](https://opam.ocaml.org/blog/opam-2-2-0/), the opam team provides an official install script. I changed it slightly so that we can install it into the cygwin directory, which appears to be necessary in our case. With this change, we also no longer need to install winget.

Pull Request resolved: https://github.com/facebook/flow/pull/9193

Test Plan: CI on GitHub

Reviewed By: gkz

Differential Revision: D60145123

Pulled By: SamChou19815

fbshipit-source-id: c062d3fa113d5652f16ad253e69a4e7445271e97