[flow][refactor][EZ] Factor out ast_extraction_utils
Summary:
For the feature of inserting inferred type as a cast, we need to go from a user selection range to an expression AST node. This is a problem already solved with the refactor feature. This diff factors out the code, so that it can be reused without depending on all of the refactor_extract_utils.

Changelog: [internal]

Reviewed By: gkz

Differential Revision: D60144103

fbshipit-source-id: af9fce48ef50cb05288acdc8e8d3a426fc066a5f