[flow][autoimport] Better support for symlinked node_modules
Summary:
To explain why the support for symlinked node_modules is bad, we need to look at the current logic:

If we have `foo/bar/baz.js` autoimporting `foo/node_modules/my-package/index.js`, we will find that they have a common prefix `foo`, and the divergent require path starts with `node_modules/my-package`, so we can directly import `my-package`. Note that the same logic will also work if the node_modules will also work if the node_modules is at the root or in `foo/bar`.

So what exactly is broken? Well symlinks break the assumption very badly. In Flow server, only absolute paths are tracked. The algorithm above only works if every path we deal with is a realpath, but will break down when a node_modules directory is symlinked to somewhere else.

In this diff, I implemented some logic to recover the fact with some best effort. If parts of the required path is in a package directory (directory that contains package.json), we redo the node module resolution algorithm to see whether the src can actually import it as node_modules. This generalized solution also makes it support yarn-workspace style node_modules setup, which I added a test.

Changelog: [ide] Flow now supports symlinked node_modules in autoimports better. If allowed by the node resolution algorithm, autoimport will insert import specifier like `my-package` or `my-package/foo` instead of a relative path import.

Reviewed By: jbrown215

Differential Revision: D60083779

fbshipit-source-id: d2523dc3d2c55a5668644a5b1fe8c84264ddaf7d