[flow] Implement non-strict eq check outside of the constraint system, in a more correct way
Summary:
For an `a == b` check, we need to distribute over unions on both sides, and for each pair of member from both unions, we run the check defined by `Flow_js_utils.equatable`. This is what the slingshot trick done by `EqT` is implicitly doing.

Sadly, this constraint system currently goes through an optimization (`union_optimization_guard`) that's completely wrong for this use case from the beginning. The optimization tries to do a subset check of LHS and RHS to decide whether to shortcut. This only makes sense for subtyping optimization, but not for equality check, where the expected behavior is to run the check on a cartesian product of all pairs.

As I am trying to move these kinds of things out of the constraint system, I see an opportunity to fix this incorrect behavior, with an even more optimized implementation.

We start from an observation that: in a relatively well-linted codebase, the only `==` that are allowed to exist is `something == null`. `Flow_js_utils.equatable`, `null` along with `void` and `mixed`, and the more general `any` and `empty`, are the types will pass the eq check if they appear on either LHS and RHS, without looking at the other half at all. Therefore, we can exploit this structure to skip potentially expensive concretization. This optimization probably covers the majority of the `==` in the codebase, so that we can let the remaining cases run on an unoptimized, but correct version.

I implemented the idea in `type_operation_utils`, by using `possible_concrete_types_for_oprators_checking` for concretization, since it has the same concretization behavior as `EqT`, except for UnionT difference (EqT differs that for optimization purposes, which as discussed above is completely wrong).

Changelog: [errors] More invalid compare errors will be consistently caught when using `==` and `!=`.

Reviewed By: mvitousek

Differential Revision:
D60211465

------------------------------------------------------------------------
(from 385d1719678d2e1dbf7ebe7527a0a70fd14fcc78)

fbshipit-source-id: bb8bb90f2e50c8ab0bc52e77198c7066555e81a9