[flow][refinement-highlighting] Initial scaffolding on Flow binary side
Summary:
In this diff, we setup initial structure necessary on the flow binary side to make the refinement highlighting feature to work. The client will send to the server `textDocument/semanticDecoration` request, which the server will respond with an array of decoration of shape `{ range, kind }`. For now, the only kind is `refined-value`, but in the future this can potentially include more semantic information that doesn't fit into any other standard LSP stuff. (e.g. fb-specific module refs)

These refined values' locations are stored in the context. For now, I start with an imprecise one, by storing all locations where refinement can possibly apply. This will be improved later, when we compute whether we actually refine in Predicate_kit.

Gating on the flow binary side is not necessary, since it's only exposed via a non-standard lsp request, so we only need to gate it on the vscode side.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D60426330

fbshipit-source-id: d47f46139f78045bd11c3ccb0984169f8c28e0b5