[flow][tuples] Allow generic bound by inexact tuple as function rest args
Summary:
Changelog: [feature] Allow generic bound by inexact tuple as function rest args

The following example now works
```
function a<T: [...]>(args: T, f: (...T) => void) {
  f(...args); // Previously an error, now OK
}
```
Previously we always constructed the rest args as an array, and an array is not a subtype of a tuple, here if the last spread is an inexact tuple we use that.

Instead of passing the element type, we pass in the full arrtype (that can be either an array with an element type, or a tuple).

Reviewed By: SamChou19815

Differential Revision: D60267613

fbshipit-source-id: 836cea459dcd50c5e821f9c2bd501bcddcbeb3df