[flow][refactor] Replace SealGenericT in predicate_kit with a simple flow
Summary:
The purpose of the setup is to ensure that `refine(GenericT(bound) -> GenericT(bound=refine(bound))`. Since there is no additional action to perform with `cond`, it's much simpler to just flow `GenericT(bound=refined_bound)` to the result tvar, instead of going through the SealGenericT indirection.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D60486604

fbshipit-source-id: 30f809fdc1125799b6a7c709d5c94a5c8e6c05f4