[flow][refactor] Replace SealGenericT in predicate_kit with a simple flow
Summary:
The purpose of the setup is to ensure that `refine(GenericT(bound) -> GenericT(bound=refine(bound))`. Since there is no additional action to perform with `cond`, it's much simpler to just flow `GenericT(bound=refined_bound)` to the result tvar, instead of going through the SealGenericT indirection.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D60486604

fbshipit-source-id: 829bc002a16a514f18cc460bb627314c8a1b3c79