[flow] Use a special kind of OpaqueT to model InternalEnforceUnionOptimizedT
Summary:
InternalEnforceUnionOptimizedT, which powers `$Flow$EnforceUnionOptimized`, is really not that special. We do a `InternalEnforceUnionOptimizedT ~> UnionT` flow and intercept in speculation, and all other interactions will result in an error.

All we need is a special norminal type that can be easily pattern matched and fail subtyping with almost everything else. We can achieve this via a special variant of opaque type. I extended opaque_id to accept a variant, extended with `InternalEnforceUnionOptimized` as a special id. In the future, it can be further extended to represent generic EvalT's that cannot be further evaluated.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision:
D60535148

------------------------------------------------------------------------
(from d52edddb0c466b83b8ddb09cd7fa71b0a84c653d)

fbshipit-source-id: ef4d200e365415736f3fc2160deddefbe0df64f4