[flow] Special-case syntactic `Object.getPrototypeof` call instead of making it a `CustomFunT`
Summary:
This diff aligns the special casing behavior with other object functions, by only special casing it in statement.ml, and give a less precise and accurate typing for it if you try to use it in other weird ways.

The ultimate motivation is to kill the final CustomFunT.

Changelog: [libdef] `Object.getPrototypeof` is still special cased, but the special-casing is only available for a syntactic `Object.getPrototypeof(...)` call. If you try to use `Object.getPrototypeof` in some other ways (e.g. `const f = Object.getPrototypeof; f(...)`), you will get a a less precise and accurate typing for it. The `Object$GetPrototypeOf` type, which was used to back the special case behavior, is removed.

Reviewed By: gkz

Differential Revision: D60603433

fbshipit-source-id: 058fc5bf72b44552e18558b9b844d780f24250bc