[flow] Setup and add tests to prepare for properly indexing `export * from '...'` for autoimports
Summary:
This diff adds some tests to assert the existing behavior of autoimports with `export * from '...'` (they are completely dropped).

As another preparation, I also make `exports.ml` produce `ReExportModule` and `ReExportModuleTypes`, but make export_service ignore them to preserve the existing behavior. Note that in the saved state `ReExportModule` and `ReExportModuleTypes` are stored, so this can be done via a rollout since the saved state will be the same in both configs.

Changelog: [internal]

Reviewed By: gkz

Differential Revision: D60545783

fbshipit-source-id: e23e0d009ac8a7a03995adfe33ed774267de8f30