[flow] Introduce flag to gate behavior of properly indexing `export * from ...`
Summary:
In the last diff in the stack, I implemented the behavior that properly indexes `export * from ...` statements. It will be a potentially risky change because it will index more and index in a different stage of a recheck, so I will gate it by a flag. This diff introduces the flag.

Changelog: [internal]

Reviewed By: gkz

Differential Revision: D60545784

fbshipit-source-id: 0b5182997f9abfbafe44faad967eb38e5905a316