[flow] Introduce flag to gate behavior of properly indexing `export * from ...`
Summary:
In the last diff in the stack, I implemented the behavior that properly indexes `export * from ...` statements. It will be a potentially risky change because it will index more and index in a different stage of a recheck, so I will gate it by a flag. This diff introduces the flag.

Changelog: [internal]

Reviewed By: gkz

Differential Revision:
D60545784

------------------------------------------------------------------------
(from 0b5182997f9abfbafe44faad967eb38e5905a316)

fbshipit-source-id: 842d23754efe9a6c3d034efa4390f8dd61b7f88b