[flow] Properly index `export * from '...'` for autoimports, behind a flag
Summary:
In this diff, we implement the proper indexing behavior for `export * from '...'`. We need to resolve the imports, and recursively include all the exports from this imported module. With this change, updating the index will need to update the index for all the modules where the signature has changed (otherwise we will get stale results due to re-exported values changes). We will experiment to see the perf impact.

Changelog: [internal]

Reviewed By: gkz

Differential Revision: D60545785

fbshipit-source-id: 0a6e60558b1f9c72a96475a1efff43feb449c740