[flow] Make all deprecated-type and untyped-type-import lint on by default
Summary:
I want to add another lint that's on by default to ban unnecessary use of internal types like `React$ElementConfig`, `React$Element` and `$EnumValue`, but found that the default lint severity we have in lintSettings.ml almost do nothing. This diff fixes it, and make `untyped-type-import` actually on by default. I also make `deprecated-type` lint on by default, since it's important to communicate to the user that some types are deprecated.

Changelog: [errors] Alldeprecated-type and untyped-type-import lint are on and at error level by default.

Reviewed By: gkz

Differential Revision: D60682181

fbshipit-source-id: 0fd25a341c3fa2e19c5e4f3bd28099f70bafc947