[flow] Lint and quickfix for use of internal Flow types
Summary:
This diff sets up the infra to ban use of internal flow types, and start to error on a subset of them. Later, I will also include `React$Element` to the list, to prepare for its eventual deprecation.

Changelog: [errors] Use of some internal types of Flow that has other identical public-facing variants (e.g. `React$ElementConfig` vs `React.ElementConfig`) is now an lint error that's enabled by default. To disable it, set `internal-type=off` in the lints section of flowconfig. For now, these types are still allowed in libdef files, but we plan to also disallow it in the future. To codemod existing code, get the latest `flow-upgrade` package and run `yarn run flow-codemod replaceReactDollarUtilityTypes`

Reviewed By: gkz

Differential Revision: D60682476

fbshipit-source-id: 28017472a9c507f783e82ebf2ed631d37e87ce33