[flow][refactor][EZ] Standardize predicate result collection
Summary:
These functions are all ported from various part of flow_js that's used to very far apart, and end up using all kinds of naming conventions for the result tvar of refinement predicate. It makes it hard to track that whether a flow here flow into an intermediate result, or the final collector tvar of the refinement result.

This diff standardizes it with a common function `flow_to_predicate_result` to collect the refinement result, and consistently name the final collector tvar `result_tvar`, to make reasoning about this code much easier.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D60625880

fbshipit-source-id: 52cf70bcb63572f6f6e707414ed12092c9072018