[flow][refactor] Use type collector for ConcretizeT
Summary:
In the next diff, I plan to use the same data structure to collect the types for predicate refinement. This diff introduces the structure and use it for `ConcretizeT` to collect all the concretized types.

Probably as a result of no longer storing types in the context that cannot be GCed and then immediate grabbing it out, we get some small perf improvement.

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D60797633

fbshipit-source-id: a116a97599dd39793cb543a58ab2291eb76fac63