[flow][render-type] Move `<: React$Node` validation of render type argument into post inference render type validation
Summary:
Render type validation is always on and cannot be disabled. In the non-generic case, the validation is already stricter than`<: React$Node`, since it requires the argument to be a union of component syntax elements. We are only relying on the subtyping check in the generic case. We can achieve the same thing by moving the validation into the post inference validation by flowing the bound of `GenericT` to `React$Node`. In this way, all the render type validation code is located together.

Changelog: [internal]

Differential Revision: D60791175

fbshipit-source-id: 7569e88382cc9b5ee961135f090fc9789386132d