[flow][render-type] Move `<: React$Node` validation of render type argument into post inference render type validation
Summary:
Render type validation is always on and cannot be disabled. In the non-generic case, the validation is already stricter than`<: React$Node`, since it requires the argument to be a union of component syntax elements. We are only relying on the subtyping check in the generic case. We can achieve the same thing by moving the validation into the post inference validation by flowing the bound of `GenericT` to `React$Node`. In this way, all the render type validation code is located together.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision:
D60791175

------------------------------------------------------------------------
(from 7569e88382cc9b5ee961135f090fc9789386132d)

fbshipit-source-id: 51c31446aacd524295f3c2f5ed01123d0ecf771e