[flow] Propagate change tracking to type_env
Summary:
In this diff, we will make use of change tracking introduced for type_filter, and make predicate_kit be able to report whether the refinement actually refines things.

This diff mostly does boring refactor to propagate the information. In the future, in order to test the check in the type checking test, we can make the refinement result be the unconcretized input, if we found that the refinement changes nothing. (Right now, if we do this, we will get error diffs, which suggests that bugs might still exist).

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision:
D60803529

------------------------------------------------------------------------
(from 630d714e467a6887c9eaef27c1f9e68c95e4958f)

fbshipit-source-id: e6fa1417d5434c5e8c7bf2b8e4dccbe23c7aa4d2