[flow] Simple last defense against possibly non-terminating concretization
Summary:
After we removed constraint caching of `~> ConcretizeT`, we start to see some recursion limit exceeded. A repro of this is added in the `rec` test suite. It's caused by a gap in bad cyclic type defense, wrt EvalT from type sig. D61055521 tries to close the gap, but the perf impact is too big.

Meanwhile, we can easily add the defense back. We don't need to restore the full constraint caching. Note that cyclic type can only come through OpenT, so we can just defend it with a seen set for a concretization operation. This set is stored as a ref in the `ConcretizeT` constraint.

Changelog: [internal]

Reviewed By: gkz

Differential Revision: D60984231

fbshipit-source-id: 9464f361e70144c2d9b378bb4e24a137a59f95a4