[flow][render-type] Model missing renders declaration as `DefaultRenders`
Summary:
In this diff, we model missing renders in component declaration as `DefaultRenders`, to replace the previous `StructuralRenders (Normal, t=React.Node)` representation.

With this change, we only need to add a few rules, the `DefaultRenders` only unwraps to `React.Node` in the general case of `l ~> DefaultRenders`. Under this change, we now have the guarantee that RHS of `RendersT ~> RendersT` check is always in a simple normalized form that only contain nominal elements or just `DefaultRenders`.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D61294566

fbshipit-source-id: 2ff10fc3e486fc87bcc104c8a8b60474f355bc76