[flow][render-type] Inline logic for TryRenderTypePromotionT
Summary:
Instead of going through the slingshot trick with TryRenderTypePromotionT, we can just inline the logic.

It's done in the try_promote_render_type_from_react_element_type helper. We first inline the logic for `TryRenderTypePromotionT` with `tried_promotion=false`, and then replace the `tried_promotion=true` case with a concretization and case analysis of `AnyT`, `RendersT` and everything else.

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D61313022

fbshipit-source-id: 57493fa873f4993a5c4d629a95d86472f5b798d3