[flow] Make cache aware of whether a write state is actually refined
Summary:
Right now, the recording of whether a write state for a given location is done behind some code that might be skipped due to cache, so we are effectively only highlighting the first read of refined values.

This diff fixes that by making whether it's actually refined part of the cache.

Changelog: [internal]

Reviewed By: gkz

Differential Revision: D61229556

fbshipit-source-id: 44fcb84fe823a3cd1e87e941b96a40825c50b304