[flow][render-type] Ban `renders* GenericT`
Summary:
Supporting this under a normalized render type world is quite tricky and involves some recursive types that's hard to get the error message in a nice way. Fortunately, there is not much use of this. It doesn't quite make sense to return `renders* T` when you only access to `T`, because it means that you need to create a list of items for the extra power, and you have to duplicate `T`, which is weird and not useful.

Therefore, let's ban it for now.

Changelog: [errors] `renders* T` where `T` is generic is now banned.

Reviewed By: jbrown215

Differential Revision:
D61336197

------------------------------------------------------------------------
(from 17c2a78e36a74719dab32f3180658d0bf53dfdcc)

fbshipit-source-id: 5be8a7ea0fabe3f547e22671dd501e47578dffe0