[flow] Fix location of symbol location within hover for VSCode
Summary:
I want to use the same logic to power the feature of hover to get refinement information. Let's fix the existing logic first.

I decide to put the logic in Loc, since the other to_string function is used to print out the error in command line, which is what detailed error feature use to jump to the right location in vscode.

So many random location standards...

xkcd_standards2

Changelog: [fix] Clicking on related symbol location on hover will jump to the right location in VSCode. Previously it incorrectly jumped to the position 1 character left of the right location.

Reviewed By: panagosg7

Differential Revision: D61496282

fbshipit-source-id: 3c5ec4242b22c6d1e5313e6603df736a4072725d