[flow] add missing attr codeaction computes edits directly
Summary:
Most code actions are implemented by computing a new AST with the necessary changes applied. We then compare this to the initial AST and compute a set of patches using the ast_differ. This approach incurs minimal burden to building new quickfixes, but it can lead to larger patches than necessary due to limitations of the differ.

Computing large patches was problematic in two ways:
1. the pretty printing machinery is not ideal. The less code we need to pretty print the better.
2. we had observed issues in the interaction of this quickfix with textDocument/linkedEditRange when the patch included the children of the element being completed, leading to garbled output.

This diff makes the "Add missing attributes" quickfix compute the patches directly, and thus the patches are minimal in size. In this case this is straightforward: the patches are insertions and are interspersed alongside the existing properties.

Compared to the previous approach this will typically lead to multiple insertion edits rather than a single replacement edit.

For example in
```
declare component Foo(
    a0: string,
    a1: string,
    a2: string,
    a3: string,
    b1: string,
    b2: string,
    c1: string,
    c2: string,
);

function test() {
    return <Foo a1={''} b1={''} c1={''}>
//             ⎸        ⎸       ⎸        ⎸
//           a0      a2,a3    b2       b3
        <div></div>
    </Foo>;
}
```
we add the attributes in the positions shown in the comment. This results to
```
return <Foo a0={a0} a1={''} a2={a2} a3={a3} b1={''} b2={b2} c1={''} c2={c2}>
```
Note that a2 and a3 will have the same location for the insertion. This is [supported by LSP](https://microsoft.github.io/language-server-protocol/specifications/lsp/3.17/specification/#textEditArray):
> If multiple inserts have the same position, the order in the array defines the order in which the inserted strings appear in the resulting text.

The approach seems to help with both issues mentioned above.

Changelog: [fix] Fixed poor interaction of "Add missing attributes" quickfix when Linked Editing Range is enabled.

Reviewed By: SamChou19815

Differential Revision: D61604033

fbshipit-source-id: 5dbbc4905e8d9f718d43fa4571895cffac10ee1b