[flow] glob recursively when rerunning tests
Summary:
The `./tool [test|record] --rerun-failed` functionality was broken for test dirs where the results.json was not in the toplevel of the project. This was the case for example for
```
./tool test newtests/lsp/code-action/quickfix
```
where results.json was under
```
/tmp/flow/tests/<some_hash>/lspzScode-action/quickfix/results.json
```
This change fixes this by globing with `'%s/**/results.json'`.

We use similar globing in `findTestSuites`.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D61602638

fbshipit-source-id: 118e1cf3dada580a8a313d540e096ff71048f2ef