[flow][vscode] Introduce flag to control whether to include refinement information in hover
Summary:
The refinement information like `this value is refined here` or `this value's refinement in invalidated here` should be exposed by hover. Eventually, we want to unconditionally expose them, but for now, we need to gate it so that we can land the change in Flow and easily see it for ourselves. I decide to do the gating in vscode so that we don't have to do a full server restart just to see these additional information.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D61603858

fbshipit-source-id: 04db7ab884768b942e2911fc417519180cfbfb45