[flow][refactor] Keep track of refining locs
Summary:
I want to provide refinement information on hover, so that the user can figure out on their own what it means when they see the green highlighting on refined values. To do that we need to track the refining locs.

Fortunately, this is already tracked in the first element of `Env_api.refinement` tuple. There are a few locations that are not ideal, but we can fix it afterwards. This diff does the refactor to make it a struct so we know what to put there, and also track it in the context by replacing the `actually_refined` bool with `actually_refined_refining_locs` `Nel.t option`

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D61603856

fbshipit-source-id: 09cf33a4c1a34a5a4870ad73fdea60d342509ae6