Add flow-dev-tools binary declaration (#9197)
Summary:
Greetings,

this will allows for using npm|yarn|pnpm link and gain an easier access to the tool.

The flow "tool", despite being unofficial, has gaining an important part in our workflow, Flow version migrations, or internal packages type changes.

Currently the only way to access and use it I'm aware of is cloning the flow repository and calling the binary by its path.
This often leads to a relative path mess such as:
```
$ ../../../../github/flow/packages/flow-dev-tools/bin/tool add-comments
```

By just declaring the binary in its `package.json` we can now `npm link` it and access it globally.
```
$ flow-dev-tools/bin/tool add-comments
```

Pull Request resolved: https://github.com/facebook/flow/pull/9197

Reviewed By: SamChou19815

Differential Revision: D61656134

fbshipit-source-id: 19d1a8eab69dccda5631d502b7303dea877997ff