[flow] Allow custom react jsx typing under a config
Summary:
D56951038 tries to simplify jsx typing, but it ends up breaking the typing of `react` repo that's making it very hard to upgrade. The difficulty comes from the fact that, things like Fragment in the react repo don't have component type, but are symbols. Currently in the react repo, it happens to type check because the `createElement` function Flow used to look up for jsx typing is any typed. We want React to keep having the latest version of Flow, so we must solve this.

An obviously solution is via global jsx pragma config, so that the react repo can keep using the loose typing through lookup of `React.createElement`. I decided against this, because our current jsx pragma impl depends on the fact that we will actually use the location of the pragma definition. It's not a problem when the pragma is defined in the same file and we can easily point to, but when the definition is in flowconfig that's not considered a file, things become much more complicated.

I go with another approach: under a config, lookup another type for jsx typing. This is similar to the [TS behavior that look up types in specific places to customize behavior](https://www.typescriptlang.org/docs/handbook/jsx.html). The implementation is mostly straightforward: under the new flag, we pass props and children directly as arguments to the new user defined type `React$CustomJSXFactory`.

Changelog: [feature] Flow now allows to configure the typing of jsx creation under the new option `react.custom_jsx_typing=true`. Under this flag, Flow will type check jsx by through your self-defined global type `React$CustomJSXFactory` type, which should be a function type that accepts arguments according to jsx spec. e.g. You can write a loose one like `type React$CustomJSXFactory = (component: any, props: any, ...children: any) => React.MixedElement`.

Reviewed By: panagosg7

Differential Revision: D61432879

fbshipit-source-id: ccca438ae73911aff0b7d6fedf906c8402a9fe16