[flow] signature help for JSX
Summary:
Adds support for LSP's Signature Help for JSX attributes.
```
component Foo(foo: number, bar?: string)) { ... }

component Bar() {
   return <Foo foo={/* here */} />;
}
```
When opening the braces after the `foo` attribute of `Foo`, there will be a tooltip with
```
foo: number
```

The approach is similar to signature help for regular function calls.

We first get the type of the component and then destructure with `ReactElementConfigType`. In most cases, this should return an object type, in which we can lookup the specific property.

This implementation also accounts for less common cases where the config of the component is:
* a union: we return multiple entries, similar to how we return multiple entries for overloaded calls.
* an intersection: we return the intersection of the intermediate results.

Changelog: [feature] Added LSP signature help support for JSX attributes. The feature is triggered with the opening "{" of a JSX attribute value.

Reviewed By: SamChou19815

Differential Revision: D59972842

fbshipit-source-id: 1caf6cfafca46719668382199f0bbbb1c0b168a1