[flow] flag to disable signature help for jsx
Summary:
Add `signature_help_for_jsx` to disable feature if necessary during rollout.

"{" is set as a trigger character in the LSP options which are hard to reset. So instead of disabling requests starting with "{" from the client, we just ignore them in the server when the flag is set to false.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D61705075

fbshipit-source-id: be19214ef025f105f990edefe9aad74b0a59deab