[flow][tests] Emit refinement information into errors under `dev_only.refinement_info_as_errors`
Summary:
This diff starts to test the refinement information feature in an automated way. We could just test this via LSP, but that's extremely painful. In this diff, I created a dev-only flag, and emit these information as errors under the flag. There are several benefits:

1. The test can be run as checking tests, which improves iteration speed.
2. The error printing already has good infra of showing inline reference, so in the error snapshot we can easily see whether the locations are in the right place.
3. The refinement type checking tests and these information snapshots are co-located in the same place.

In this diff, I enables it for the refinement test suite. I inspected every single one, and added a few TODO comments, where I think there is an obvious bug.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D61671191

fbshipit-source-id: f9f146ea0d8e161ece96a485bf26da163ef39c7a