[flow] Initial implementation to show refinement invalidation on hover
Summary:
In this diff, I sketched out an initial implementation of showing that a value is invalidated on hover. It's intentionally not comprehensive, and not very precise right now, but I think it will work well in practice.

In name_resolver, we will augment the heap_refinement state in `heap_refinements` from a map of `Val.t` to map of `(Val.t, Refinement_invalidation.t) result`. On some occasions where we currently havoc heap refinements, we will still havoc them, but we will store the invalidation information in place of the result values. Currently, this applies to a function-call like thing invalidating all heap refinement, a named property assignment invalidating all heap refinements with the name, and indexed property assignment invalidating all heap refinements.

The invalidation information mostly behaves like a missing refinement entry as before, but we keep this optimistic state around, so that by the time we read from a value that might be refined, we can know whether we can read a refined value if we never invalidate.

Note that this scheme doesn't apply universally. e.g. right now, we also use havoc to reset to initialization state from an undeclared state. In these cases, it doesn't make sense to store things to the snapshot. We also keep the old behavior of completely forgetting refinements when we decide that we will invalidate the refinement on the base variable as well. Therefore, this extra information will only be shown on hover for heap refinements in limited cases, but it should cover the vast majority of confusing cases, where it appears to the user that everything is constant but refinement just disappears out of nowhere.

There are still a lot we can improve here. e.g. In the hover, we can point to the invalidation location.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D61645181

fbshipit-source-id: 45fbb72eb411c705693ba3a3854e01417d441419