[flow] Fix spurious refinement highlighting on switch statement
Summary:
The problem is supposed to be fixed by D60988628, but broken by D61229556 again. The issue is that if a value going through multiple actual refinements, then even if we recognize that it's not a source-level read at the end, we will still say it's refined because it's OR-merged with refined results.

The real solution is to do the `val_kind <> Env_api.Internal` check right before the final moment before we add the refinement information into the context.

Changelog: [internal]

Reviewed By: gkz

Differential Revision: D61675233

fbshipit-source-id: 9d6ebe42f74f732fa35f51603f62262f7bb4dc53