[flow] Fix spurious refinement highlighting on prop tests
Summary:
Right now, we will refinement highlight the final `foo` read in

```
declare const foo: {bar: string | null}
if (foo.bar) {
    foo
}
```

This happens because we generate a `PropTruthyP('bar')` for `foo`. Then for each concretized `foo.bar` union type, we run it against the `guard` function. Right now, if any of the prop get filtered out, we will report `foo` as being refined. However, this is incorrect, because `foo` is only filtered to empty if all the members of union of `foo.bar` get filtered to empty.

This diff fixes the issue.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D61689085

fbshipit-source-id: b9203ee4ac0f4be3d91f38dec338acac2bea08de