[flow][autocomplete] show optional information in member label
Summary:
Before this diff when autocompleting the following
```
function test(obj: { foo?: number }) {
    obj.f|
}
```
we would show `foo: number | void`.

This diff changes this to show the fact that `foo` is optional: `foo?: number | void`.

I decided to keep the `| void` part since it is more indicative of the type that can be assigned to `foo`.

This change also applies to:
```
const obj: T = { fo| };
<Foo fo| />;
```

Changelog: [fix] labels of autocomplete on members will indicate if the property is optional with a "?"

Reviewed By: SamChou19815

Differential Revision: D61737097

fbshipit-source-id: f88c08e8fc3bac445e5305b832eb8f2c7f7a520e