[flow] Use the full expression location as hover loc for refinement and invalidation information
Summary:
Right now, we will provide refinement and invalidation information on hover based on the exact location match of the type-at-pos location and the map in the context. This works fine for variables. e.g.

```
if (foo) {
//  vvv refinement info loc, also type-at-pos loc
    foo
 //  ^ cursor
}
```

but it's not fine for props:

```
if (foo.bar) {

//  vvvvvvv refinement info loc
//      vvv type-at-pos loc
    foo.bar
 //      ^ cursor
}
```

I think we should make the hover for refinement information work for the entire expression, since the entire expression is highlighted. For refinement invalidation, we should also make it cover the entire expression, since given a type error, the entire expression is red-underlined.

This diff implements the location cover logic in type_info_service, and revert the implementation around hover_loc in D61645181.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D61690677

fbshipit-source-id: 0ccd46ee9e5a0ab24133de1590d080acb5c3f488