[flow] Add tests to demonstrate invalidation info behavior difference in conditional vs unconditional abrupt completions
Summary:
In conditional abrupt completions, we will correctly show the invalidation status, while for unconditional ones below, we do not.

I think the cause is that the env empty in the name_resolver is modeled incorrectly: it should represent an unreachable state, but in reality it causes all the heap refinements to be gone. Anyway, this diff adds the test to show the behavior difference first.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D61864170

fbshipit-source-id: dc63be2ab34f2d979e405dd36f4353a05518f649