[flow] Remove special case for react element shorthand syntax
Summary:
Let's stop the bleeding first so that people don't just copy paste what's in the hover and end up with tons of exact react element requirements.

I do have to add special support for nominal elements by normalizing it as `typeof C`, similar to how we treat classes.

Changelog: [ide] Elements of component syntax components  (e.g. typeof `<A />` where A is in component syntax) will no longer be shown in hover with shorthand syntax like `A`. Instead, it will be shown as `React$Element<typeof A>`

Reviewed By: panagosg7

Differential Revision: D61955053

fbshipit-source-id: eecfd5e25df50bd251ec8d0bc7e4a275e2c8f140